+++ psakeFile.ps1
@@ -1,21 +1,22 @@
 properties {
     $PSBPreference.General.SrcRootDir = 'src'
     $PSBPreference.General.ModuleName = 'Datto.DBPool.Refresh'
-    $PSBPreference.Build.OutDir = $projectRoot
+    $PSBPreference.Build.OutDir = "$projectRoot"
     # Set this to $true to create a module with a monolithic PSM1
     $PSBPreference.Build.CompileModule = $true
     $PSBPreference.Build.CopyDirectories = @('scripts')
-    $PSBPreference.Build.CompileScriptHeader = [System.Environment]::NewLine + '#Region'
-    $PSBPreference.Build.CompileScriptFooter = "#EndRegion"
+    $PSBPreference.Build.CompileScriptHeader = '#Region' + [System.Environment]::NewLine
+    $PSBPreference.Build.CompileScriptFooter = [System.Environment]::NewLine + '#EndRegion'
     $PSBPreference.Help.DefaultLocale = 'en-US'
     $PSBPreference.Test.OutputFile = 'out/testResults.xml'
     $PSBPreference.Test.ImportModule = $true
     $PSBPreference.Docs.RootDir = "./docs/generated"
 }
 
-task Default -depends Test
+Task Default -Depends Test
 
-Task UpdateFunctionsToExport -depends StageFiles {
+# Custom task to update functions export
+Task UpdateFunctionsToExport -Depends StageFiles {
     Write-Host 'Updating module manifest to update FunctionsToExport section'
 
     $exportDirectory = 'Public', 'Private'
@@ -30,21 +31,43 @@ Task UpdateFunctionsToExport -depends StageFiles {
 
     if ($allFunctions) {
         $modulePath_Built = Join-Path -Path $env:BHBuildOutput -ChildPath $( $env:BHProjectName + '.psd1' )
-
         # Update FunctionsToExport with all functions in the module
         Update-Metadata -Path $modulePath_Built -PropertyName FunctionsToExport -Value $allFunctions.BaseName
-
         Write-Host 'Module manifest updated successfully'
     }
 }
 
+# Custom task to remove nested modules - run after StageFiles but before GenerateMarkdown
+Task RemoveNestedModules -Depends UpdateFunctionsToExport {
+    $modulePath_Built = Join-Path -Path $env:BHBuildOutput -ChildPath $( $env:BHProjectName + '.psd1' )
 
-Task Build -FromModule PowerShellBuild -depends @('StageFiles', 'UpdateFunctionsToExport', 'BuildHelp')
+    Write-Host 'Updating module manifest to remove NestedModules section'
+    # Replace & comment out NestedModules from built module
+    Update-Metadata -Path $modulePath_Built -PropertyName NestedModules -Value @()
+    (Get-Content -Path $modulePath_Built -Raw) -replace 'NestedModules = @\(\)', '# NestedModules = @()' | Set-Content -Path $modulePath_Built
+    Write-Host 'Module manifest updated successfully'
+}
 
-task Test -FromModule PowerShellBuild -minimumVersion '0.6.1'
+# Override GenerateMarkdown to depend on RemoveNestedModules
+Task GenerateMarkdown -FromModule PowerShellBuild -Depends RemoveNestedModules
 
-task PublishDocs -depends Build {
-    exec {
-        docker run -v "$($psake.build_script_dir)`:/docs" -e 'CI=true' --entrypoint 'sh' squidfunk/mkdocs-material:9 -c 'pip install -r requirements.txt && mkdocs gh-deploy --force'
+# Override Build to include your custom dependencies
+Task Build -FromModule PowerShellBuild -Depends @('GenerateMarkdown', 'BuildHelp')
+
+# Test task inherits from PowerShellBuild and will depend on Build
+Task Test -FromModule PowerShellBuild -MinimumVersion '0.6.1' -Depends Build
+
+Task PublishDocs -Depends Build {
+    $env:GITHUB_TOKEN = $env:GITHUB_TOKEN ?? ''
+    $env:GITHUB_REPOSITORY = $env:GITHUB_REPOSITORY ?? ''
+    $env:GITHUB_ACTOR = $env:GITHUB_ACTOR ?? ''
+
+    Exec {
+        docker run -v "$($psake.build_script_dir)`:/docs" `
+            -e 'CI=true' `
+            -e "GITHUB_TOKEN=$env:GITHUB_TOKEN" `
+            -e "GITHUB_REPOSITORY=$env:GITHUB_REPOSITORY" `
+            -e "GITHUB_ACTOR=$env:GITHUB_ACTOR" `
+            --entrypoint 'sh' squidfunk/mkdocs-material:9@sha256:3bba0a99bc6e635bb8e53f379d32ab9cecb554adee9cc8f59a347f93ecf82f3b -c 'pip install -r requirements.txt && mkdocs gh-deploy --force'
     }
 }